chore(e2e): move shared consts utils and types
