+++ e2e/scripts/types.ts
@@ -1,0 +1,5 @@
+export interface Token {
+  symbol: string
+  address: string // Mento expects address to be in checksum format, or else it won't find the trading pair
+  decimals: number
+}
